// fig/types/cheat.ts
  export interface Subcommand extends BaseSuggestion {
    /**
     * The exact name of the subcommand. It is important to get this right for parsing purposes.
     *
     * @example
     * For `npm install`, the subcommand `install` would have "name: install"
     */
    name: SingleOrArray<string>;
  
    /**
     * A list of subcommands for this spec.
     * Subcommands can be nested recursively.
     */
    subcommands?: Subcommand[];
  
    /**
     * A list of [option objects](./option) for this subcommand.
     */
    options?: Option[];
  
    /**
     * An array of args or a single arg.
     *
     * If a subcommand takes an argument, please at least include an empty Arg Object
     * (e.g. `{}`). If you don't, Fig will assume the subcommand does not take an argument and we will present the wrong suggestions.
     *
     * @example:
     * git push takes two arguments. The most basic representation of this is
     * `args: [{}, {}]`
     */
  
    args?: SingleOrArray<Arg>;
  
    /**
     * A list of Suggestion objects that are appended to a specific subcommand. These are often shortcuts.
     *
     * @example:
     * `commit -m '{cursor}'` is a shortcut for git
     */
    additionalSuggestions?: Suggestion[] | string[];
    /**
     * Allows Fig to load another completion spec in the `~/.fig/autocomplete` folder.
     * Specify the spec name without `js`.
     *
     * @example
     * `aws-s3` refers to the `~/.fig/autocomplete/aws-s3` spec.
     *
     * When is this used? The aws spec is so large that it is slow to load. It needs to be
     * broken up into a separate spec for each subcommand.
     *
     * If your CLI tool takes another CLI command (e.g. time , builtin) or a script
     * (e.g. python, node) and you would like Fig to continue to provide completions for this
     * script, see `isCommand` and `isScript` in @link https://fig.io/docs/reference/arg | Arg}.
     */
    loadSpec?: string;
    /**
     * Dynamically generate a completion spec to be merged in at the same level as the current subcommand. This is useful when a CLI is generated dynamically.
     * This function takes two params:
     * 1. Tokens: an array of strings (the tokens the user has typed)
     * 2. executeShellCommand: a function that takes a string as input. It executes this string as a shell command on the user's device from the same current working directory as their terminal. It outputs a text blob. It is also async.
     *
     * generateSpec outputs a completion spec object.
     *
     * @example
     * The python spec uses generateSpec to insert the django-admin spec if django manage.py exists.
     */
    generateSpec?: (tokens?: string[], executeShellCommand?: ExecuteShellCommandFunction) => Promise<Spec>;
  
    /**
     * These flags allow customization of how Fig parses tokens from the
     * command line.
     *
     * Currently, the only parser option that exists is flagsArePosixNoncompliant. This
     * allows options to have multiple characters even if they only have one hyphen when set to `true`.
     *
     *
     * @example
     * `-work` from the go spec is parsed as a single flag when `parserDirectives.flagsArePosixNoncompliant` is set to true. Normally, this would be chained and parsed as `-w -o -r -k` if `flagsArePosixNoncompliant` is not set to true.
     */
    parserDirectives?: {
      flagsArePosixNoncompliant?: boolean;
    };
  }
  export type Spec = Subcommand;
  
  // Execute shell command function inside generators
  export type ExecuteShellCommandFunction = (commandToExecute: string) => Promise<string>;
  
  export interface BaseSuggestion {
    /**
     * The string that is displayed in the UI for a given suggestion. This overrides the name property.
     *
     * @example
     * The npm CLI has a subcommand called `install`. If we wanted
     * to display some custom text like `Install an NPM package 📦` we would set
     * `name: "install"` and `displayName: "Install an NPM package 📦"`
     */
    displayName?: string;
    /**
     * The value that's inserted into the terminal when a user presses enter/tab or clicks on a menu item. You can use `\n` to insert a newline or `\b` to insert a backspace.
     * You can also optionally specify {cursor} in the string and Fig will automatically place the cursor there after insert.
     * The default is the name prop.
     *
     * @example
     * For the `git commit` subcommand, the `-m` option has an insert value of `-m '{cursor}'`
     */
    insertValue?: string;
    /**
     * The text that gets rendered at the bottom of the autocomplete box (or the side if you hit ⌘i)
     */
    description?: string;
    /**
     * The icon that is rendered is based on the type. Icons
     * can be a 1 character string, a URL, or Fig's [icon protocol](https://fig.io/docs/reference/suggestion/icon-api) (fig://) which lets you generate
     * colorful and fun systems icons.
     *
     * @example
     * `A`, `😊`
     * `https://www.herokucdn.com/favicon.ico`
     * `fig://icon?type=file`
     */
    icon?: string;
    /**
     * Specifies whether the suggestion is "dangerous". If true, Fig will not enable
     * its "insert and run" functionality (when Fig has the red insert icon).
     * This will make it harder for a user to accidentally run a dangerous command.
     *
     * @example
     * This is used in the `rm` spec.
     */
    isDangerous?: boolean;
  
    /**
     * The priority between 0-100 for a given suggestion determines its ranking in the Fig popup. A higher ranked priority will be listed first. The default priority is 50.
     * If a given suggestion has a priority between 50-75 inclusive AND has been selected by the user before, the priority will be replaced with 75 + the timestamp of when that suggestion was selected as a decimal.
     *
     * If a given suggestion has a priority outside of 50-75 AND has been selected by the user before, the priority will be increased by the timestamp of when that suggestion was selected as a decimal.
     *
     *
     * @example
     * If you want your suggestions to always be at the top order regardless of whether they have been selected before or not, rank them 76 or above.
     *
     * If you want your suggestions to always be at the bottom regardless of whether they have been selected before or not, rank them 49 or below
     */
    priority?: number;
    /**
     * Specifies whether a suggestion should be hidden from results. Fig will only show it if the user types the exact name.
     *
     * @example
     * The "-" suggestion is hidden in the `cd` spec. You will only see it if you type `cd -`
     */
    hidden?: boolean;
  }
  export type TemplateStrings = "filepaths" | "folders";
  
  export type Template = TemplateStrings | TemplateStrings[];
  
  export type SuggestionType = "folder" | "file" | "arg" | "subcommand" | "option" | "special" | "shortcut";
  
  export interface Suggestion extends BaseSuggestion {
    /**
     * The text that’s rendered in each row of Fig's popup. displayName will override this.
     *  The `name` props of suggestion, subcommand, options, and args objects are all different. It's important to read them all carefully.
     *
     */
    name?: SingleOrArray<string>;
    /**
     * The type of suggestion, one of `folder`, `file`, `arg`, `subcommand`, `option`,
     * `special`.
     */
    type?: SuggestionType;
  }
  
  export type SingleOrArray<T> = T | T[];